refactor: DispatcherHandler 제거
- DispatcherHandler의 역할이 모호하여 제거했습니다.
