+++ src/main/java/handler/RequestHandler.java
@@ -1,54 +1,34 @@
 package handler;
 
+import java.io.*;
+import java.net.Socket;
+
 import http.Request;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import util.RequestParser;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.util.LinkedHashMap;
-import java.util.Map;
+public class RequestHandler implements Runnable {
+    private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
 
-public class RequestHandler {
-    private InputStream in;
+    private Socket connection;
 
-    public RequestHandler(InputStream in) {
-        this.in = in;
+    public RequestHandler(Socket connectionSocket) {
+        this.connection = connectionSocket;
     }
 
-    public Request handleRequest() throws IOException {
-        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(in));
-        Map<String, String> requestLine = parseRequestLine(bufferedReader);
-        Map<String, String> headers = parseRequestHeaders(bufferedReader);
-        return new Request(requestLine, headers);
-    }
+    public void run() {
+        logger.debug("New Client Connect! Connected IP : {}, Port : {}", connection.getInetAddress(),
+                connection.getPort());
 
-    private Map<String, String> parseRequestLine(BufferedReader bufferedReader) throws IOException {
-        Map<String, String> requestLine = new LinkedHashMap<>();
-        String[] requestLineParts = bufferedReader.readLine().split(" ");
-        requestLine.put("method", requestLineParts[0]);
-        requestLine.put("path", requestLineParts[1]);
-        requestLine.put("protocol", requestLineParts[2]);
-        return requestLine;
-    }
-
-    private Map<String, String> parseRequestHeaders(BufferedReader bufferedReader) throws IOException {
-        Map<String, String> headers = new LinkedHashMap<>();
-
-        String line;
-
-        while ((line = bufferedReader.readLine()) != null) {
-
-            if (line.isEmpty()) break;
-
-            String[] headerParts = parseLine(line);
-            headers.put(headerParts[0], headerParts[1]);
+        try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
+            Request request = RequestParser.parseRequest(in);
+            request.print();
+            ResponseHandler responseHandler = new ResponseHandler(out, request.getRequestLine("path"));
+            responseHandler.sendResponse();
+        } catch (IOException e) {
+            logger.error(e.getMessage());
         }
-
-        return headers;
     }
 
-    private String[] parseLine(String line) {
-        return line.split(":");
-    }
 }
